finish all loot prefabs + loot ID tracking
by default loot will be given ID of -1. ID should always be -1 unless loot is associated with a quest, in which case it should have an ID matching the ID of the quest. This helps if two people ask for help to find their missing object when both objects are of the same type.